Fix: Implement cube matrix in correct updateGeometry function
- Remove duplicate matrix code from rt-rendering.js updateGeometry()
- Keep matrix support only in rt-init.js updateGeometry() (the actual function being called)
- Remove unused RTMatrix import from rt-rendering.js
- Remove debug console.log statements from event handlers
- Clean up matrix slider and rotate checkbox event handlers
- Matrix functionality now working correctly with dynamic import
