+++ src/geometry/modules/rt-rendering.js
@@ -12,7 +12,6 @@
 
 import { Quadray } from "./rt-math.js";
 import { Polyhedra } from "./rt-polyhedra.js";
-import { RTMatrix } from "./rt-matrix.js";
 
 /**
  * Initialize THREE.js scene and return rendering context
@@ -589,35 +588,10 @@ export function initScene(THREE) {
     const scale = tetEdge / (2 * Math.sqrt(2)); // Convert tet edge to halfSize
     const opacity = parseFloat(document.getElementById("opacitySlider").value);
 
-    // Cube (Blue) - with Matrix support
+    // Cube (Blue)
     if (document.getElementById("showCube").checked) {
-      // Get matrix settings
-      const matrixSize = parseInt(
-        document.getElementById("matrixSizeSlider")?.value || "1"
-      );
-      const rotate45 = document.getElementById("matrixRotate45")?.checked || false;
-
-      // Clear existing cube group
-      while (cubeGroup.children.length > 0) {
-        cubeGroup.remove(cubeGroup.children[0]);
-      }
-
-      if (matrixSize > 1) {
-        // Use matrix generator for N×N array
-        const cubeMatrix = RTMatrix.createCubeMatrix(
-          matrixSize,
-          scale,
-          rotate45,
-          opacity,
-          0x4a9eff,
-          THREE
-        );
-        cubeGroup.add(cubeMatrix);
-      } else {
-        // Single cube (original behavior)
-        const cube = Polyhedra.cube(scale);
-        renderPolyhedron(cubeGroup, cube, 0x4a9eff, opacity);
-      }
+      const cube = Polyhedra.cube(scale);
+      renderPolyhedron(cubeGroup, cube, 0x4a9eff, opacity);
       cubeGroup.visible = true;
     } else {
       cubeGroup.visible = false;
